+++ lib/models.dart
@@ -2,7 +2,8 @@ export 'src/models/author.dart' show Author;
 export 'src/models/book.dart' show Book;
 export 'src/models/chapter.dart' show Chapter;
 export 'src/models/chapter_list.dart' show ChapterList;
-export 'src/models/content.dart' show Content, MarkdownContent;
+export 'src/models/content.dart'
+    show Content, MarkdownContent, IndexableMarkdownContent;
 export 'src/models/heading.dart' show Heading;
 export 'src/models/index.dart' show Index;
 export 'src/models/reference.dart' show Reference;