+++ lib/src/models/content.dart
@@ -1,5 +1,6 @@
 import 'package:powerbook/mixins.dart';
-export 'content/markdown_content.dart' show MarkdownContent;
+export 'content/markdown_content.dart'
+    show MarkdownContent, IndexableMarkdownContent;
 
 /// A generic class that represents a single piece of content
 abstract class Content with Renderable {