+++ lib/src/models/content/markdown_content.dart
@@ -1,4 +1,5 @@
 import 'package:markdown/markdown.dart';
+import 'package:powerbook/mixins.dart';
 import 'package:powerbook/models.dart';
 
 /// A `Content` type that returns the rendered results from a markdown string
@@ -12,6 +13,36 @@ class MarkdownContent extends Content {
   String toHtml() {
     return markdownToHtml(markdown, blockSyntaxes: blockSyntaxes)
         .replaceAll(RegExp(r"(\n+)"), "");
+  }
+}
+
+class IndexableMarkdownContent extends MarkdownContent with Indexable {
+  final List<Heading> headings = [];
+
+  IndexableMarkdownContent({required String markdown})
+      : super(markdown: markdown);
+
+  @override
+  String getId() {
+    return headings.first.getId();
+  }
+
+  @override
+  String getTitle() {
+    return headings.first.getTitle();
+  }
+
+  @override
+  List<Indexable> getChildren() {
+    return headings.sublist(1);
+  }
+
+  @override
+  String toHtml() {
+    return markdownToHtml(markdown, blockSyntaxes: [
+      CollectHeadersForIndexable(headings),
+      ...super.blockSyntaxes
+    ]).replaceAll(RegExp(r"(\n+)"), "");
   }
 }
 
@@ -31,5 +62,25 @@ class HeaderSyntaxWithHeading extends BlockSyntax {
     var level = match[1]!.length;
     var contents = match[2]!.trim();
     return Text(Heading(title: contents, headingLevel: level).toHtml());
+  }
+}
+
+class CollectHeadersForIndexable extends HeaderSyntaxWithHeading {
+  final List<Heading> headings;
+
+  @override
+  RegExp get pattern => _headerPattern;
+
+  const CollectHeadersForIndexable(this.headings);
+
+  @override
+  Node parse(BlockParser parser) {
+    var match = pattern.firstMatch(parser.current)!;
+    parser.advance();
+    var level = match[1]!.length;
+    var contents = match[2]!.trim();
+    var heading = Heading(title: contents, headingLevel: level);
+    headings.add(heading);
+    return Text(heading.toHtml());
   }
 }